refactor: move FX API types into shared module
