+++ frontend/src/api/dashboard/index.ts
@@ -2,9 +2,6 @@ export type {
   BreakdownCategoryKind,
   CategoryBreakdownEntry,
   CreditWidgetResponse,
-  FxRateIssue,
-  FxState,
-  FxStatus,
   MonthlyIncomeExpense,
   NetWorthWidgetResponse,
   RecentActivityWidgetResponse,
