+++ frontend/src/api/shared/fx.ts
@@ -1,0 +1,11 @@
+export type FxState = 'none' | 'complete' | 'incomplete' | 'unavailable';
+
+export interface FxRateIssue {
+  base: string;
+  quote: string;
+}
+
+export interface FxStatus {
+  state: FxState;
+  missing_pairs: FxRateIssue[];
+}
