+++ frontend/src/api/transactions/types.ts
@@ -1,4 +1,4 @@
-import type { FxStatus } from '@/api/dashboard';
+import type { FxStatus } from '@/api/shared/fx';
 
 export interface Transaction {
   id: string;
