+++ frontend/src/budgets/utils/fxTooltipMessages.ts
@@ -1,4 +1,4 @@
-import type { FxStatus } from '@/api/dashboard'
+import type { FxStatus } from '@/api/shared/fx'
 
 export function getBudgetUtilizationFxStatusMessage(fxStatus: FxStatus) {
   switch (fxStatus.state) {
