+++ frontend/src/dashboard/utils/fxTooltipMessages.ts
@@ -1,4 +1,4 @@
-import type { FxStatus } from '@/api/dashboard'
+import type { FxStatus } from '@/api/shared/fx'
 
 export function getNetWorthFxStatusMessage(fxStatus: FxStatus) {
   switch (fxStatus.state) {
