+++ frontend/src/insights/components/IncomeExpenseBreakdownCard.tsx
@@ -13,7 +13,7 @@ import {
   PieChart,
   ResponsiveContainer,
 } from 'recharts'
-import type { FxStatus } from '@/api/dashboard'
+import type { FxStatus } from '@/api/shared/fx'
 import IconTooltip from '@/components/IconTooltip'
 import { getIncomeExpenseBreakdownFxStatusMessage } from '@/insights/utils/fxTooltipMessages'
 import { formatCurrency } from '@/utils/formatCurrency'
