+++ frontend/src/insights/utils/fxTooltipMessages.ts
@@ -1,4 +1,4 @@
-import type { FxStatus } from '@/api/dashboard'
+import type { FxStatus } from '@/api/shared/fx'
 
 export function getPeriodIncomeExpenseFxStatusMessage(fxStatus: FxStatus) {
   switch (fxStatus.state) {
