+++ frontend/src/transactions/components/topBand/MostExpensiveTransactionsPanel.tsx
@@ -1,5 +1,5 @@
 import { AnimatePresence, motion } from 'motion/react'
-import type { FxStatus } from '@/api/dashboard'
+import type { FxStatus } from '@/api/shared/fx'
 import type { OutlierTransaction } from '@/api/transactions'
 import IconTooltip from '@/components/IconTooltip'
 import { formatMissingFxPairs, getFxStatusTone } from '@/dashboard/utils/fxStatus'
